Implemented all functionality. The next step is styling.

+++ src/app/Http/Controllers/MessageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreMessageRequest;
 use App\Http\Requests\UpdateMessageRequest;
 use App\Models\Message;
+use App\Models\Tickets;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
@@ -18,6 +19,10 @@ class MessageController extends Controller
      */
     public function index(): JsonResponse
     {
+        if (!Auth::user()->is_support) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
         $messages = Message::all();
 
         return response()->json($messages, 200);
@@ -31,14 +36,15 @@ class MessageController extends Controller
      */
     public function store(StoreMessageRequest $request): JsonResponse
     {
-        // check whether the ticket is related to the user
-//        $ticket = Tickets::find($request->ticket_id);
-//        if ($ticket->status === 'resolved') {
-//            return response()->json(['error' => 'Ticket is resolved'], 403);
-//        }
-//        if ($ticket->user_id !== Auth::user()->id || !Auth::user()->is_suport) {
-//            return response()->json(['error' => 'Unauthorized'], 401);
-//        }
+        $ticket = Tickets::find($request->ticket_id);
+        if ($ticket->status === 'resolved') {
+            return response()->json(['error' => 'Ticket is resolved'], 403);
+        }
+
+        if ($ticket->user_id !== Auth::user()->id && !Auth::user()->is_support) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
         $requestParams = $request->all();
         $requestParams += ['user_id' => Auth::user()->id];
 
@@ -51,10 +57,14 @@ class MessageController extends Controller
      * Display the specified resource.
      *
      * @param Message $message
-     * @return Message
+     * @return Message|null
      */
-    public function show(Message $message): Message
+    public function show(Message $message): ?Message
     {
+        if ($message->user_id !== Auth::user()->id && !Auth::user()->is_support) {
+            return null;
+        }
+
         return $message;
     }
 
@@ -67,6 +77,10 @@ class MessageController extends Controller
      */
     public function update(UpdateMessageRequest $request, Message $message): JsonResponse
     {
+        if ($message->user_id !== Auth::user()->id && !Auth::user()->is_support) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
         $message->update($request->all());
         return response()->json($message, 200);
     }
@@ -79,10 +93,10 @@ class MessageController extends Controller
      */
     public function destroy(Message $message): JsonResponse
     {
+        if ($message->user_id !== Auth::user()->id && !Auth::user()->is_support) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
 
-        // check if user has the specified rights to delete the ticket or is the author of the ticket.
-        // if everything ok - go on.
-        // otherwise response with an error
         $message->delete();
 
         return response()->json(null, 204);
